# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Finance Management API",
        "description": "Backend API for the personal finance management app.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL with port for local development"
        }
    ],
    "paths": {
...
        "/api/v1/chat/query": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Chat with AI about transaction behavior",
                "description": "Chat with AI about user's transaction behavior and spending patterns.\n\nThe AI will analyze the user's recent transactions and provide insights\nbased on their query about spending habits, patterns, or financial behavior.",
                "operationId": "chat_query_api_v1_chat_query_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatQueryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatQueryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/chat/context": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Chat context data",
                "description": "Return summarized financial data for the AI chat context.",
                "operationId": "chat_context_api_v1_chat_context_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
        "/api/v1/chat/history": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Chat history",
                "operationId": "chat_history_api_v1_chat_history_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChatQueryRequest": {
                "properties": {
                    "query": {
                        "type": "string",
                        "title": "Query"
                    }
                },
                "type": "object",
                "required": [
                    "query"
                ],
                "title": "ChatQueryRequest"
            },
            "ChatQueryResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response"
                    },
                    "context_used": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Context Used"
                    }
                },
                "type": "object",
                "required": [
                    "response",
                    "context_used"
                ],
                "title": "ChatQueryResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}